update documentation + remove session_refresher.ex
